convert react and angular to use default and gallery icon views
