+++ angular/.storybook/config.js
@@ -4,7 +4,7 @@ import { withA11y } from '@storybook/addon-a11y';
 import sparkTheme from "../../storybook-utilities/storybook-theming/storybook-spark-theme";
 import '../src/polyfills';
 import '!style-loader!css-loader!sass-loader!../../storybook-utilities/storybook-theming/font-loader.scss';
-import '../../storybook-utilities/icon-loader';
+import '../../storybook-utilities/icon-utilities/icon-loader';
 import { setCompodocJson, extractProps } from '@storybook/addon-docs/angular';
 import docJson from '../documentation.json';
 import { DocsContainer } from '@storybook/addon-docs/blocks';
