+++ html/.storybook/config.js
@@ -6,7 +6,7 @@ import { withA11y } from '@storybook/addon-a11y';
 import sparkTheme from '../../storybook-utilities/storybook-theming/storybook-spark-theme';
 import 'iframe-resizer/js/iframeResizer.contentWindow.min';
 import '!style-loader!css-loader!sass-loader!../../storybook-utilities/storybook-theming/font-loader.scss';
-import '../../storybook-utilities/icon-loader';
+import '../../storybook-utilities/icon-utilities/icon-loader';
 import { DocsContainer } from '@storybook/addon-docs/blocks';
 import { SprkTable } from '@sparkdesignsystem/spark-react';
 import { configClassModifierJsonProcessor } from '../../storybook-utilities/configClassModifierJsonProcessor';
