+++ react/src/components/icons/SprkIcon.stories.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import SprkIcon from './SprkIcon';
 import { markdownDocumentationLinkBuilder } from '../../../../storybook-utilities/markdownDocumentationLinkBuilder';
+import { getIcons, attachIcons, createIconExample } from "../../../../storybook-utilities/icon-utilities/icon-name-util";
 
 export default {
   title: 'Components/Icons',
@@ -14,923 +15,34 @@ export default {
   },
 };
 
-export const access = () => (
+export const defaultStory = () => (
   <SprkIcon iconName="access" additionalClasses="sprk-c-Icon--xl" />
 );
 
-export const amountMoney = () => (
-  <SprkIcon iconName="amount-money" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const application = () => (
-  <SprkIcon iconName="application" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const applyNow = () => (
-  <SprkIcon iconName="apply-now" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const autoLoans = () => (
-  <SprkIcon iconName="auto-loans" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const baths = () => (
-  <SprkIcon iconName="baths" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const beds = () => (
-  <SprkIcon iconName="beds" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const bell = () => (
-  <SprkIcon iconName="bell" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const bellFilled = () => (
-  <SprkIcon
-    iconName="bell-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const bellFilledSmall = () => (
-  <SprkIcon
-    iconName="bell-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const blog = () => (
-  <SprkIcon iconName="blog" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const calculator = () => (
-  <SprkIcon iconName="calculator" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const calendar = () => (
-  <SprkIcon iconName="calendar" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const callTeamMember = () => (
-  <SprkIcon iconName="call-team-member" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const camera = () => (
-  <SprkIcon iconName="camera" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const cashBack = () => (
-  <SprkIcon iconName="cash-back" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chatOnline = () => (
-  <SprkIcon iconName="chat-online" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chatOnlineFilled = () => (
-  <SprkIcon
-    iconName="chat-online-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const chatOnlineFilledSmall = () => (
-  <SprkIcon
-    iconName="chat-online-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const checkMark = () => (
-  <SprkIcon iconName="check-mark" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const checkMarkFilled = () => (
-  <SprkIcon
-    iconName="check-mark-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const checkMarkFilledSmall = () => (
-  <SprkIcon
-    iconName="check-mark-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const chevronDown = () => (
-  <SprkIcon iconName="chevron-down" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chevronDownCircle = () => (
-  <SprkIcon
-    iconName="chevron-down-circle"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronDownCircleTwoColor = () => (
-  <SprkIcon
-    iconName="chevron-down-circle-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronDownFilled = () => (
-  <SprkIcon
-    iconName="chevron-down-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const chevronDownFilledSmall = () => (
-  <SprkIcon
-    iconName="chevron-down-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const chevronLeft = () => (
-  <SprkIcon iconName="chevron-left" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chevronLeftCircle = () => (
-  <SprkIcon
-    iconName="chevron-left-circle"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronLeftCircleTwoColor = () => (
-  <SprkIcon
-    iconName="chevron-left-circle-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronLeftFilled = () => (
-  <SprkIcon
-    iconName="chevron-left-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const chevronLeftFilledSmall = () => (
-  <SprkIcon
-    iconName="chevron-left-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const chevronRight = () => (
-  <SprkIcon iconName="chevron-right" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chevronRightCircle = () => (
-  <SprkIcon
-    iconName="chevron-right-circle"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronRightCircleTwoColor = () => (
-  <SprkIcon
-    iconName="chevron-right-circle-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronRightFilled = () => (
-  <SprkIcon
-    iconName="chevron-right-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const chevronRightFilledSmall = () => (
-  <SprkIcon
-    iconName="chevron-right-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const chevronUp = () => (
-  <SprkIcon iconName="chevron-up" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chevronUpCircle = () => (
-  <SprkIcon iconName="chevron-up-circle" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const chevronUpCircleTwoColor = () => (
-  <SprkIcon
-    iconName="chevron-up-circle-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const chevronUpFilled = () => (
-  <SprkIcon
-    iconName="chevron-up-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const chevronUpFilledSmall = () => (
-  <SprkIcon
-    iconName="chevron-up-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const clock = () => (
-  <SprkIcon iconName="clock" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const clockFilled = () => (
-  <SprkIcon
-    iconName="clock-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const clockFilledSmall = () => (
-  <SprkIcon
-    iconName="clock-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const close = () => (
-  <SprkIcon iconName="close" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const closeCircle = () => (
-  <SprkIcon iconName="close-circle" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const closeFilled = () => (
-  <SprkIcon
-    iconName="close-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const closeFilledSmall = () => (
-  <SprkIcon
-    iconName="close-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const cloud = () => (
-  <SprkIcon iconName="cloud" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const cloudFilled = () => (
-  <SprkIcon
-    iconName="cloud-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const cloudFilledSmall = () => (
-  <SprkIcon
-    iconName="cloud-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const communication = () => (
-  <SprkIcon iconName="communication" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const communicationFilled = () => (
-  <SprkIcon
-    iconName="communication-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const communicationFilledSmall = () => (
-  <SprkIcon
-    iconName="communication-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const condo = () => (
-  <SprkIcon iconName="condo" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const cooling = () => (
-  <SprkIcon iconName="cooling" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const credit = () => (
-  <SprkIcon iconName="credit" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const creditCardPayoff = () => (
-  <SprkIcon iconName="credit-card-payoff" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const dedicatedTeam = () => (
-  <SprkIcon iconName="dedicated-team" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const desktopComputer = () => (
-  <SprkIcon iconName="desktop-computer" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const documentReview = () => (
-  <SprkIcon iconName="document-review" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const documentReviewFilled = () => (
-  <SprkIcon
-    iconName="document-review-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const documentReviewFilledSmall = () => (
-  <SprkIcon
-    iconName="document-review-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const documents = () => (
-  <SprkIcon iconName="documents" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const dollar = () => (
-  <SprkIcon
-    iconName="dollar"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const download = () => (
-  <SprkIcon iconName="download" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const downloadFilled = () => (
-  <SprkIcon
-    iconName="download-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const downloadFilledSmall = () => (
-  <SprkIcon
-    iconName="download-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const editable = () => (
-  <SprkIcon iconName="editable" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const educationalResources = () => (
-  <SprkIcon
-    iconName="educational-resources"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const email = () => (
-  <SprkIcon iconName="email" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const emailFilled = () => (
-  <SprkIcon
-    iconName="email-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const emailFilledSmall = () => (
-  <SprkIcon
-    iconName="email-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const exclamation = () => (
-  <SprkIcon iconName="exclamation" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const exclamationCredit = () => (
-  <SprkIcon iconName="exclamation-credit" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const exclamationFilled = () => (
-  <SprkIcon
-    iconName="exclamation-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const exclamationFilledSmall = () => (
-  <SprkIcon
-    iconName="exclamation-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const facebook = () => (
-  <SprkIcon iconName="facebook" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const facebookTwoColor = () => (
-  <SprkIcon iconName="facebook-two-color" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const feedback = () => (
-  <SprkIcon iconName="feedback" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const feedbackFilled = () => (
-  <SprkIcon
-    iconName="feedback-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const feedbackFilledSmall = () => (
-  <SprkIcon
-    iconName="feedback-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const findAgent = () => (
-  <SprkIcon iconName="find-agent" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const free = () => (
-  <SprkIcon iconName="free" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const heat = () => (
-  <SprkIcon iconName="heat" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const home = () => (
-  <SprkIcon iconName="home" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const homeImprovement = () => (
-  <SprkIcon iconName="home-improvement" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const homeShopping = () => (
-  <SprkIcon iconName="home-shopping" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const homeSize = () => (
-  <SprkIcon iconName="home-size" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const information = () => (
-  <SprkIcon iconName="information" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const informationFilled = () => (
-  <SprkIcon
-    iconName="information-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const informationFilledSmall = () => (
-  <SprkIcon
-    iconName="information-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const instagram = () => (
-  <SprkIcon iconName="instagram" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const instagramTwoColor = () => (
-  <SprkIcon
-    iconName="instagram-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const land = () => (
-  <SprkIcon iconName="land" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const landline = () => (
-  <SprkIcon iconName="landline" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const landlineFilled = () => (
-  <SprkIcon
-    iconName="landline-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const landlineFilledSmall = () => (
-  <SprkIcon
-    iconName="landline-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const laptop = () => (
-  <SprkIcon iconName="laptop" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const largePurchase = () => (
-  <SprkIcon iconName="large-purchase" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const linkedin = () => (
-  <SprkIcon iconName="linkedin" additionalClasses="sprk-c-Icon--xl" />
-);
-
-linkedin.story = {
-  name: 'LinkedIn',
+defaultStory.story = {
+  name: 'Default',
 };
 
-export const linkedinTwoColor = () => (
-  <SprkIcon iconName="linkedin-two-color" additionalClasses="sprk-c-Icon--xl" />
-);
+export const gallery = () => {
+  useEffect(() => {
+    attachIcons(getIcons());
+  }, []);
 
-linkedinTwoColor.story = {
-  name: 'LinkedIn Two Color',
+  window.addEventListener('sprk-icons-loaded', () => {
+    attachIcons(getIcons());
+  });
+
+  return (<ul className="
+    sbdocs-gallery-container
+    sprk-o-Stack
+    sprk-o-Stack--split@xs"></ul>);
 };
 
-export const lock = () => (
-  <SprkIcon iconName="lock" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const medicalExpenses = () => (
-  <SprkIcon iconName="medical-expenses" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const menu = () => (
-  <SprkIcon iconName="menu" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const mls = () => (
-  <SprkIcon iconName="mls" additionalClasses="sprk-c-Icon--xl" />
-);
-
-mls.story = {
-  name: 'MLS',
+gallery.story = {
+  name: 'Gallery',
+  parameters: {
+    docs: { iframeHeight: 1000 },
+  },
 };
 
-export const mobile = () => (
-  <SprkIcon iconName="mobile" additionalClasses="sprk-c-Icon--xl" />
-);
 
-export const mortgage = () => (
-  <SprkIcon iconName="mortgage" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const multiFamily = () => (
-  <SprkIcon iconName="multi-family" additionalClasses="sprk-c-Icon--xl" />
-);
-
-multiFamily.story = {
-  name: 'Multi-Family',
-};
-
-export const multiplePages = () => (
-  <SprkIcon iconName="multiple-pages" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const oneTimeUpdate = () => (
-  <SprkIcon iconName="one-time-update" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const organize = () => (
-  <SprkIcon iconName="organize" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const organizeFilled = () => (
-  <SprkIcon
-    iconName="organize-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const organizeFilledSmall = () => (
-  <SprkIcon
-    iconName="organize-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const ourFamily = () => (
-  <SprkIcon iconName="our-family" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const parking = () => (
-  <SprkIcon iconName="parking" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const payment = () => (
-  <SprkIcon iconName="payment" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const paymentCenter = () => (
-  <SprkIcon iconName="payment-center" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const percent = () => (
-  <SprkIcon
-    iconName="percent"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const pinterest = () => (
-  <SprkIcon iconName="pinterest" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const pinterestTwoColor = () => (
-  <SprkIcon
-    iconName="pinterest-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const play = () => (
-  <SprkIcon iconName="play" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const plus = () => (
-  <SprkIcon iconName="plus" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const plusFilled = () => (
-  <SprkIcon
-    iconName="plus-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const plusFilledSmall = () => (
-  <SprkIcon
-    iconName="plus-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const prequalified = () => (
-  <SprkIcon iconName="prequalified" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const prequalifiedFilled = () => (
-  <SprkIcon
-    iconName="prequalified-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const prequalifiedFilledSmall = () => (
-  <SprkIcon
-    iconName="prequalified-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const pricePerSqFt = () => (
-  <SprkIcon iconName="price-per-sq-ft" additionalClasses="sprk-c-Icon--xl" />
-);
-
-pricePerSqFt.story = {
-  name: 'Price Per SQ FT',
-};
-
-export const print = () => (
-  <SprkIcon iconName="print" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const question = () => (
-  <SprkIcon iconName="question" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const questionFilled = () => (
-  <SprkIcon
-    iconName="question-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const questionFilledSmall = () => (
-  <SprkIcon
-    iconName="question-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const rateShield = () => (
-  <SprkIcon iconName="rate-shield" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const rateShieldFilled = () => (
-  <SprkIcon
-    iconName="rate-shield-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const rateShieldFilledSmall = () => (
-  <SprkIcon
-    iconName="rate-shield-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const rates = () => (
-  <SprkIcon iconName="rates" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const recurringUpdate = () => (
-  <SprkIcon iconName="recurring-update" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const refinance = () => (
-  <SprkIcon iconName="refinance" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const reviews = () => (
-  <SprkIcon iconName="reviews" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const rocket = () => (
-  <SprkIcon iconName="rocket" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const rocketDocTwoColor = () => (
-  <SprkIcon
-    iconName="rocket-doc-two-color"
-    additionalClasses="sprk-c-Icon--xl"
-  />
-);
-
-export const rotatePage = () => (
-  <SprkIcon iconName="rotate-page" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const sameDay = () => (
-  <SprkIcon iconName="same-day" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const savings = () => (
-  <SprkIcon iconName="savings" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const schools = () => (
-  <SprkIcon iconName="schools" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const search = () => (
-  <SprkIcon iconName="search" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const secureProcess = () => (
-  <SprkIcon iconName="secure-process" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const settings = () => (
-  <SprkIcon iconName="settings" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const settingsFilled = () => (
-  <SprkIcon
-    iconName="settings-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const settingsFilledSmall = () => (
-  <SprkIcon
-    iconName="settings-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const single = () => (
-  <SprkIcon iconName="single" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const smallBusiness = () => (
-  <SprkIcon iconName="small-business" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const snapchat = () => (
-  <SprkIcon iconName="snapchat" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const snapchatTwoColor = () => (
-  <SprkIcon iconName="snapchat-two-color" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const technology = () => (
-  <SprkIcon iconName="technology" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const thumbsUp = () => (
-  <SprkIcon iconName="thumbs-up" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const townhouse = () => (
-  <SprkIcon iconName="townhouse" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const trash = () => (
-  <SprkIcon iconName="trash" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const trashFilled = () => (
-  <SprkIcon
-    iconName="trash-filled"
-    additionalClasses="sprk-c-Icon--xl sprk-c-Icon--filled"
-  />
-);
-
-export const trashFilledSmall = () => (
-  <SprkIcon
-    iconName="trash-filled-small"
-    additionalClasses="sprk-c-Icon--filled"
-  />
-);
-
-export const twitter = () => (
-  <SprkIcon iconName="twitter" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const twitterTwoColor = () => (
-  <SprkIcon iconName="twitter-two-color" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const unlock = () => (
-  <SprkIcon iconName="unlock" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const update = () => (
-  <SprkIcon iconName="update" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const upload = () => (
-  <SprkIcon iconName="upload" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const user = () => (
-  <SprkIcon iconName="user" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const userAccount = () => (
-  <SprkIcon iconName="user-account" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const verifiedApproval = () => (
-  <SprkIcon iconName="verified-approval" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const vip = () => (
-  <SprkIcon iconName="vip" additionalClasses="sprk-c-Icon--xl" />
-);
-
-vip.story = {
-  name: 'VIP',
-};
-
-export const wifi = () => (
-  <SprkIcon iconName="wifi" additionalClasses="sprk-c-Icon--xl" />
-);
-
-wifi.story = {
-  name: 'Wi-Fi',
-};
-
-export const writtenCheck = () => (
-  <SprkIcon iconName="written-check" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const yearBuilt = () => (
-  <SprkIcon iconName="year-built" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const youtube = () => (
-  <SprkIcon iconName="youtube" additionalClasses="sprk-c-Icon--xl" />
-);
-
-export const youtubeTwoColor = () => (
-  <SprkIcon iconName="youtube-two-color" additionalClasses="sprk-c-Icon--xl" />
-);
