+++ storybook-utilities/icon-utilities/icon-loader.js
